+++ src/moduleSettingsStore.js
@@ -1,18 +1,24 @@
+let goosemodScope = {};
+
+export const setThisScope = (scope) => {
+  goosemodScope = scope;
+};
+
 export const saveModuleSettings = async () => {
-  //globalThis.logger.debug('settings', 'Saving module settings...');
+  //goosemodScope.logger.debug('settings', 'Saving module settings...');
 
   let settings = JSON.parse(localStorage.getItem('goosemodModules')) || {};
 
-  for (let p in globalThis.modules) {
-    if (globalThis.modules.hasOwnProperty(p)) {
-      settings[p] = await (globalThis.modules[p].getSettings || (async () => []))();
+  for (let p in goosemodScope.modules) {
+    if (goosemodScope.modules.hasOwnProperty(p)) {
+      settings[p] = await (goosemodScope.modules[p].getSettings || (async () => []))();
     }
   }
 
   if (JSON.stringify(JSON.parse(localStorage.getItem('goosemodModules'))) !== JSON.stringify(settings)) {
     localStorage.setItem('goosemodModules', JSON.stringify(settings));
 
-    globalThis.showToast('Settings saved');
+    goosemodScope.showToast('Settings saved');
   }
 
   //console.log(settings);
@@ -33,20 +39,19 @@ export const clearSettings = () => {
 export const loadSavedModuleSetting = async (moduleName) => {
   let settings = JSON.parse(localStorage.getItem('goosemodModules'));
 
-  await (globalThis.modules[moduleName].loadSettings || (async () => []))(settings[moduleName]);
+  await (goosemodScope.modules[moduleName].loadSettings || (async () => []))(settings[moduleName]);
 };
 
 export const loadSavedModuleSettings = async () => {
-  //globalThis.logger.debug('settings', 'Loading module settings...');
+  //goosemodScope.logger.debug('settings', 'Loading module settings...');
 
   let settings = JSON.parse(localStorage.getItem('goosemodModules'));
 
   if (!settings) return;
 
-  for (let p in globalThis.modules) {
-    if (globalThis.modules.hasOwnProperty(p) && settings.hasOwnProperty(p)) {
-      console.log(p, globalThis.modules[p].loadSettings, settings[p]);
-      await (globalThis.modules[p].loadSettings || (async () => []))(settings[p]);
+  for (let p in goosemodScope.modules) {
+    if (goosemodScope.modules.hasOwnProperty(p) && settings.hasOwnProperty(p)) {
+      await (goosemodScope.modules[p].loadSettings || (async () => []))(settings[p]);
     }
   }